requirements.txt
py-trello==0.11.3
python-dotenv
pyperclip
